pyproject.toml
[tool.poetry]
name = "llm-cookbook"
version = "0.1.0"
description = ""
authors = ["He Fei <droxer.he@gmail.com>"]
readme = "README.md"
packages = [{include = "llm-cookbook"}]


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
priority = "primary"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.4"
pandas = "^2.2.1"
matplotlib = "^3.8.3"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"
langchain-community = "^0.3.0"
pydantic = "^2.7.3"
neo4j = "^5.22.0"
pyautogen = "^0.2.32"
transformers = "^4.42.4"
torch = "^2.3.0"
sentence-transformers = "^3.0.1"
einops = "^0.8.0"
llama-cpp-python = "^0.2.83"
httpx = {extras = ["socks"], version = "^0.27.0"}
graphrag = "^0.2.1"
icecream = "^2.1.3"
langchain-experimental = "^0.3.2"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.1"
openai = "^1.55.0"
numexpr = "^2.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
